+++ src/commonMain/kotlin/com/hoc081098/flowext/cast.kt
@@ -53,3 +53,9 @@ public inline fun <reified T : Any> Flow<T?>.castNotNull(): Flow<T> = map { it!!
  */
 @Suppress("NOTHING_TO_INLINE")
 public inline fun <T> Flow<T>.castNullable(): Flow<T?> = this
+
+/**
+ * Adapt this `Flow<*>` to be a `Flow<R?>`.
+ */
+@Suppress("NOTHING_TO_INLINE")
+public inline fun <reified R> Flow<*>.safeCast(): Flow<R?> = map { it as? R }